% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesreg-package.R
\docType{package}
\name{bayesreg-package}
\alias{bayesreg-package}
\alias{bayesreg}
\title{The 'bayesreg' package.}
\description{
Performs Bayesian regularization for linear regression models.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.4. http://mc-stan.org
}
