% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_est.R
\name{plot_est}
\alias{plot_est}
\title{Plot results for Bayesian regularized regression models}
\usage{
plot_est(fit, est = c("mean", "median"), CI = 0.95, npar = 50,
  names = NULL)
}
\arguments{
\item{fit}{A named list with one or more objects of class `stanfit` returned by `stan_reg_lm`.}

\item{est}{A character indicating the posterior estimate to plot. Options are: posterior mean or median.}

\item{CI}{Numeric value specifying the probability \eqn{p \in (0,1)}{p (0 < p < 1)} indicating the desired
credibility interval to plot. The default is 0.95.}

\item{npar}{Numeric value or sequence specifying the number of parameters to include in the plot. When a numeric
value is given, all parameters are plotted with a maximum of npar per plot, thus returning multiple plots. When a
sequence is given, only the parameters given by the sequence are plotted in one plot.}

\item{names}{A character vector of parameter names to include in the plot. Make sure that the names are ordered the
same way as the columns of the input matrix X used in the `stan_reg_lm` function.}
}
\value{
A list with ggplot objects.
}
\description{
Function to plot point estimates and credibility intervals for models fit using `stan_reg_lm`.
Results from multiple fit objects can be plotted for comparison.
}
