% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_lm.R
\name{select_lm}
\alias{select_lm}
\title{Variable selection for Bayesian regularized regression models}
\usage{
select_lm(object, X, prob = c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9))
}
\arguments{
\item{object}{An object of class `stanfit` returned by `stan_reg_lm`.}

\item{X}{Numeric matrix[N, p] of input values used to fit the model including column names.}

\item{prob}{Numeric vector with probabilities \eqn{p \in (0,1)}{p (0 < p < 1)} indicating the desired
probability mass to include in the intervals. The default is to report variable selection based on all 
   intervals ranging from 10\% to 90\%, i.e., prob = c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9).}
}
\value{
Logical matrix with variables in the rows and CIs in the columns returning TRUE if the CI covers 0, and FALSE otherwise.
}
\description{
Function to perform variable selection on models fit using `stan_reg_lm`. Variable selection is based on the
credibility interval criterion, excluding predictors for which the marginal credibility interval covers 0.
}
