% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_reg_lm.R
\name{stan_reg_lm}
\alias{stan_reg_lm}
\title{Bayesian regularized linear regression with Stan}
\usage{
stan_reg_lm(X, y, N_train, prior = c("ridge", "student", "lasso",
  "elasticNet", "hyperlasso", "horseshoe", "regularizedHorseshoe", "mixture"),
  missing = c(TRUE, FALSE), df = 0.5, p0 = NA, global_scale = 1,
  global_df = 1, local_df = 1, slab_scale = 1, slab_df = 1,
  hyperprior_mix = c("Bernoulli", "uniform"), ...)
}
\arguments{
\item{X}{Numeric matrix[N, p] of input values. The input values get automatically standardized in the function.}

\item{y}{Numeric vector[N] of output values.}

\item{N_train}{Size of the training set. First part of the data will be used for training.}

\item{prior}{Shrinkage prior to use for regularization. Current options are: ridge, student, lasso,
elasticNet, hyperlasso, horseshoe, regularizedHorseshoe, and mixture.}

\item{missing}{Logical value indicating whether the data contains missing values. Currently only supported in combination with the ridge prior.}

\item{df}{Degrees of freedom for the hyperlasso. Defaults to 0.5, but larger values can lead to less divergent transitions.}

\item{p0}{Prior information on the number of relevant predictors. Used to determine the global scale parameter in the regularized horseshoe.}

\item{global_scale}{Scale parameter for the t-distribution on the global shrinkage parameter in the regularized horseshoe. Defaults to 1, but better is to provide p0 if prior info is available.}

\item{global_df}{Degrees of freedom for the t-distribution on the global shrinkage parameter in the regularized horseshoe. Defaults to 1.}

\item{local_df}{Degrees of freedom for the t-distribution on the local shrinkage parameter in the regularized horseshoe. Defaults to 1.}

\item{slab_scale}{Scale parameter for the t-distribution shrinking the large coefficients. Defaults to 1.}

\item{slab_df}{Degrees of freedom for the t-distribution shrinking the large coefficients. Defaults to 1.}

\item{hyperprior_mix}{Hyperprior for the mixing probabilities in the normal mixture prior. Options are: uniform or Bernoulli.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`.
}
\description{
Function to run regularized linear regression models using Stan. Allows the user to specify which 
shrinkage prior to use for the regression coefficients.
}
