% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_lm.R
\name{summary_lm}
\alias{summary_lm}
\title{Summary function for Bayesian regularized regression models}
\usage{
summary_lm(object, CI = 0.95, prob = NULL)
}
\arguments{
\item{object}{An object of class `stanfit` returned by `stan_reg_lm`.}

\item{CI}{Numeric value specifying the probability \eqn{p \in (0,1)}{p (0 < p < 1)} indicating the desired
probability mass to include in the reported marginal credibility interval. The default is 0.95.}
}
\value{
Numeric matrix including the posterior mean, median, standard deviation and credibility interval
for the intercept (`int`), regression coefficients, and residual standard deviation (`sigma`).
}
\description{
Function to return a summary of results for models fit using `stan_reg_lm`.
}
